+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.23.0
 ## Added
 * Support for partition default route-domains; set it to an existing route-domain with a custom FAST template
+* bigip-fast-templates: Added support for Service Discovery
 
 ## Fixed
 
@@ -12,7 +13,7 @@
 * Class Function Documentation and an option to create the associaited HTML or Markdown files 
 
 ## Fixed
-* bigip-fast-templates: uploading zipped Template Set files works in GUI (was broken in v1.20.0 and v1.21.0)
+* gui: uploading zipped Template Set files works in GUI (was broken in v1.20.0 and v1.21.0)
 * gui: Fix displaying template parameters that are not required and have no default value
 
 ## Changed